requirements.txt
python-docx==0.8.11
docxtpl==0.16.2
pandas==1.3.5
numpy==1.21.5
matplotlib==3.5.1
python-dateutil==2.8.2
argparse==1.4.0
beautifulsoup4==4.12.3
requests==2.31.0
selenium==4.19.0
webdriver-manager==4.0.1
fake-useragent==1.3.0